Keep Mailchimp in sync with model updates.
Now, when models (tickets and users) are updated, the Mailchimp
subscriber list is updated as well. So, when a first name, last name, or
email changes in a ticket, the relevant fields are updated in the
subscriber list.

This was done by adding a Sidekiq worker called
`mailchimp_unsubscriber'; hence, the previous subscriber worker is
renamed to `mailchimp_subscriber'. This rename forces a change in the
tickets controller.

The updates are done via ActiveRecord callbacks. Using methods in
ActiveRecord::Dirty, Mailchimp subscribers are deleted after an update,
and are added again after a save.