Add "add user" function to admin page.
Admins can now add users. I've added both `new' and `create' actions to
the admin user controller, as well as added a view, which is basically
the same as the Devise `new registration' view. Routes were unchanged,
since there was already a path for `new_admin_user', as well as one
handling a post to `admin_users'.