Add `add ticket' to user admin page.
Add ability to add a ticket from the admin user edit page. Add nested
routes to handle creating a new ticket for a certain user -- previously,
though a route was defined, it didn't contain an id, so posting to the
`admin_ticket_create' path was difficult.

Added `new' and `create' actions to the tickets controller, create a new
ticket admin page, and add a button to add tickets in the admin user
edit page.