Move scrollspy script to body of home page.
Previously, the scrollspy refresh script was located in the head of the
landing page. Since scrollspy actually needs to be refreshed every time
the page is loaded, and since Turbolinks only loads the content inside
the body, the scrollspy wouldn't be refreshed. I've moved it to the
body, to force a refresh.