Change scrollspy refresh javascript.
Previously, the scrollspy refresh was only called on page:load, which
wouldn't always work. So, for example, (going to home -> another page ->
hard refreshing the other page -> going back to home) would cause
scrollspy to stop working.

Now, scrollspy is also refreshed on a document.ready, which fixes that
problem. See SO:
http://stackoverflow.com/questions/16754227/re-rendering-specific-js-file-with-turbolinks-enabled