Add terms and conditions to tickets page.
Adds a T&C at the top of the tickets page. It currently links to `#' --
this will be changed when our terms and conditions are published.

Also adds a 4 pixel `padding-top' to the form on the tickets page.
Without it, the space between the paragraph and the button is less than
the space between the two paragraphs.